Move util functions
