+++ homeassistant/components/tradfri/const.py
@@ -36,3 +36,4 @@ PLATFORMS = [
     Platform.SWITCH,
 ]
 TIMEOUT_API = 30
+ATTR_MAX_FAN_STEPS = 49
