+++ homeassistant/components/tradfri/fan.py
@@ -17,7 +17,15 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .base_class import TradfriBaseDevice
-from .const import ATTR_AUTO, CONF_GATEWAY_ID, DEVICES, DOMAIN, KEY_API
+from .const import (
+    ATTR_AUTO,
+    ATTR_MAX_FAN_STEPS,
+    CONF_GATEWAY_ID,
+    DEVICES,
+    DOMAIN,
+    KEY_API,
+)
+from .utils import _from_fan_speed, _from_percentage
 
 
 async def async_setup_entry(
@@ -36,23 +44,6 @@ async def async_setup_entry(
         for dev in devices
         if dev.has_air_purifier_control
     )
-
-
-def _from_percentage(percentage: int) -> int:
-    """Convert percent to a value that the Tradfri API understands."""
-    if percentage < 20:
-        # The device cannot be set to speed 5 (10%), so we should turn off the device
-        # for any value below 20
-        return 0
-
-    nearest_10: int = round(percentage / 10) * 10  # Round to nearest multiple of 10
-    return round(nearest_10 / 100 * 50)
-
-
-def _from_fan_speed(fan_speed: int) -> int:
-    """Convert the Tradfri API fan speed to a percentage value."""
-    nearest_10: int = round(fan_speed / 10) * 10  # Round to nearest multiple of 10
-    return round(nearest_10 / 50 * 100)
 
 
 class TradfriAirPurifierFan(TradfriBaseDevice, FanEntity):
@@ -80,24 +71,19 @@ class TradfriAirPurifierFan(TradfriBaseDevice, FanEntity):
 
         These are the steps:
         0 = Off
-        10 = Min
-        15
-        20
-        25
-        30
-        35
-        40
-        45
+        1 = Preset: Auto mode
+        2 = Min
+        ... with step size 1
         50 = Max
         """
-        return 10
+        return ATTR_MAX_FAN_STEPS
 
     @property
     def is_on(self) -> bool:
         """Return true if switch is on."""
         if not self._device_data:
             return False
-        return cast(bool, self._device_data.mode)
+        return cast(bool, self._device_data.state)
 
     @property
     def preset_modes(self) -> list[str] | None:
@@ -121,7 +107,7 @@ class TradfriAirPurifierFan(TradfriBaseDevice, FanEntity):
         if not self._device_data:
             return None
 
-        if self._device_data.mode == ATTR_AUTO:
+        if self._device_data.is_auto_mode:
             return ATTR_AUTO
 
         return None
@@ -133,7 +119,8 @@ class TradfriAirPurifierFan(TradfriBaseDevice, FanEntity):
 
         if not preset_mode == ATTR_AUTO:
             raise ValueError("Preset must be 'Auto'.")
-        await self._api(self._device_control.set_mode(1))
+
+        await self._api(self._device_control.turn_on_auto_mode())
 
     async def async_turn_on(
         self,
@@ -158,13 +145,13 @@ class TradfriAirPurifierFan(TradfriBaseDevice, FanEntity):
         if not self._device_control:
             return
 
-        await self._api(self._device_control.set_mode(_from_percentage(percentage)))
+        await self._api(self._device_control.set_fan_speed(_from_fan_speed(percentage)))
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Turn off the fan."""
         if not self._device_control:
             return
-        await self._api(self._device_control.set_mode(0))
+        await self._api(self._device_control.turn_off())
 
     def _refresh(self, device: Command, write_ha: bool = True) -> None:
         """Refresh the purifier data."""
